pyproject.toml
[tool.poetry]
name = "financial-accounting"
version = "0.1.0"
description = "App for managing your expenses and debts!"
authors = ["artemkka <artiom.preatca@ebs-integrator.com>"]
readme = "README.md"
packages = [{include = "financial_accounting"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
djangorestframework = "^3.14.0"
python-dotenv = "^0.21.0"
django-jazzmin = "^2.6.0"
flake8 = "^5.0.4"
drf-yasg = "^1.21.4"
black = "^22.10.0"
pre-commit = "^2.20.0"
djangorestframework-simplejwt = "^5.2.2"
psycopg2 = "^2.9.5"
psycopg2-binary = "^2.9.5"
django-filter = "^22.1"
faker = "^15.3.4"
coverage = "^6.5.0"

[tool.black]
line-length = 120
target-version = ["py310"]


[tool.vulture]
min_confidence = 65
paths = ["apps"]
exclude = [
    "**/fixtures/*",
    "**/tests/*",
    "**/migrations/*",
    "users/models.py",
    "expenses/views.py",
    "debts/views.py",
    "incomes/views.py",
    "users/views.py",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
